Robotwars - move classes into their own folders, and general clean up
