+++ Notes.md
@@ -1,0 +1,10 @@
+Implementation Notes
+====================
+
+The projects have been developed in visual studio 2017 community edition. The following folders have been added:
+
+* RobotWars -  holds the robot wars solution.
+* SupermarketKata - holds the supermarked kata solution.
+
+
+Both projects are c# and were developed in a pure TDD style where the design emerged as I was implementing the tests.  The guard statments that were put in are not covered by tests that is purely from habit where I tend to cover the happy and sad paths with tests but not the defensive programming constructs.  I tried to stick to the 2 hr time slots but probably slightly exceeded it on both projects I checked in at aout the two hour mark.
