+++ RobotWars/RobotWars/RobotWars.csproj
@@ -40,6 +40,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Arena.cs" />
+    <Compile Include="Position.cs" />
+    <Compile Include="Robot.cs" />
     <Compile Include="RobotWar.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
